requirements.txt
python-dotenv==1.0.1
langchain==0.2.2
langchain-openai==0.1.8
langchain-community==0.2.3
chromadb==0.5.0
openai==1.31.1
unstructured==0.14.4
unstructured[md]==0.14.4
